requirements.txt
Flask==2.3.2
ultralytics==8.3.52
torch==2.0.0
clip @ git+https://github.com/openai/CLIP.git
Werkzeug==2.3.4
Pillow==9.5.0
